-lynx: Updated timer handling,most of the games that froze at the intro screen go in-game now [huygens]
git-svn-id: svn://messdev.no-ip.org/mess@13497 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/includes/lynx.h
@@ -74,7 +74,7 @@ typedef struct
 	UINT8	bakup;
 	UINT8	cntrl1;
 	UINT8	cntrl2;
-	int		counter;
+	UINT8	counter;
 	emu_timer	*timer;
 	int		timer_active;
 } LYNX_TIMER;
